pyproject.toml
[project]
name = "container"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "docker>=7.1.0",
    "gitpython>=3.1.45",
]

[[tool.uv.index]]
url = "https://mirrors.tuna.tsinghua.edu.cn/pypi/web/simple"
default = true

[dependency-groups]
dev = [
    "mkdocs>=1.6.1",
]
